+++ CHANGELOG.md
@@ -2,11 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
-## [3.0.2] 2020-04-24
+## [3.0.3] 2020-04-30
 
 ### Added
 
 - Universal node: new commands adjust/get.volume, player.adjust.volume,  set/get.mute, player.set/get.mute  
+
+## [3.0.2] 2020-04-23
+
+### Changed
+
+- Control Player: bug fix serialnum
+
+- Helper, error codes: new error handling
 
 ## [3.0.1] 2020-04-18